Initial implementation of Amberol
Amberol is a music player, so let's give it a music player UI that
consists of:

- the currently playing song, with title, artist, album, and cover art
- the current position and song duration
- a set of buttons to control playback
- a queue with all the songs loaded up

Amberol should be able to load a single song and put it in the queue,
and it should also be able to go through a directory and load all the
music files it encounters (without recursing).

Amberol goes through the queue automatically, but also has skip
functions to skip to the previous/next song; there's also a repeat mode
control that allows Amberol to repeat the current song or the whole
queue. The queue view can also control the playback by skipping to the
song on activation.

There are still missing bits, here and there; for instance:

- no shuffle
- no indicator in the queue for the currently playing song

And, more importantly, there is no application icon.