Sort folder contents before queueing them
The gio::FileEnumerator::enumerate_children() method yields children of
a folder in no particular order. We do assume that the contents of a
music folder have been ordered in some way by the its owner, so it's
better to use the URI as the sorting key. Even if the contents haven't
been ordered, relying on simple alphabetical order is going to yield
better results than random.