Experiment with UI recoloring
We use the dominant color of the cover art image to determine the
background and foreground colors of our main window. Depending on
the background color's relative luminance we either use Adwaita's
dark or light foreground color for the text.

The actual recoloring happens through a custom CSS fragment that defines
symbolic colors to be used by the application's window.

If the current song has no cover art, we remove the style class and fall
back to whatever default color is provided by the current stylesheet.