Use message passing in the AudioPlayer implementation
Instead of updating the state and queue from within signals emitted by
the GstPlayer and MprisPlayer objects, we use message passing using a
pair of sender/receiver objects created by glib::MainContext::channel()
in the AudioPlayer itself.

In order to use message passing we need to split out the AudioPlayer
into different ancillary objects:

 - the PlayerState object, which holds the current state of the
   playback, and uses GObject properties to notify the UI
 - the Queue object, which holds the gio::ListStore that contains
   all the Song instances, and is used to move to the next, previous,
   or given songs; the Queue object also uses GObject properties to
   notify the UI
 - the GstBackend object, which wraps GstPlayer, and sends messages
   to the AudioPlayer whenever GStreamer tells us that something has
   changed
 - the MprisController object, which wraps the MPRIS interface, and
   sends messages to the AudioPlayer whenever an MPRIS event happens

The AudioPlayer API is the main entry point for all player operations;
the UI uses the state and queue objects to update itself.

The main advantage of this new design is that we don't need to
complicate the lifetime and memory management of the AudioPlayer
instance for the benefit of the various closures, since the only
object of interest being passed is a glib::Sender. The Window
now keeps the AudioPlayer instance inside an Rc, and we don't need
to contort ourselves with borrows.

Fixes: #6