Use the appropriate color format for the texture data
The format used by GdkMemoryTexture depends on the endianness.

The C API has `GDK_MEMORY_DEFAULT`, but the symbol is not
introspectable, so it's not available to bindings. This means we need to
do the endianness detection ourselves, and figure out which ColorFormat
value we should pass to color_thief::get_palette().

Fixes: #7