Add a specialised cover art widget
The cover must be fixed at 128×128 pixels, which means GtkPicture cannot
really be used, as it will change its size if the aspect ratio is not 1.

Additionally, we are kind of fudging the line by setting the overflow
property on an existing class, which it should only be set on custom
widgets.