style: Use the first three colors for the recoloring gradients
Instead of skipping the second and jumping to the third.

It was mostly meant as an experiment, but now that the palette is
sampled correctly, we can use the expected behaviour.

In the future we might want to see if the colors are close to each
other, and skip them, but it requires some additional thought on the
allowed range.