ui: Use an AdwFlap to handle the playlist
Instead of a revealer, we can use a flap to display the playlist view.
In "landscape mode", it works exactly the same, but in "portrait mode"
it will be overlaid on top of the song details and playback controls.

This allows us to make the UI fully resizable, while retaining a
mobile-aware layout.

Fixes: #28