Make the waveform view more legible
Increase the size of the peak lines and the spacing between them;
draw the new cursor position when hovering; and make sure to
visually distinguish the currently played part of the song from the
new cursor position.

The code also needs some decent comments on how it works, and how
to change it in the future.

Fixes: #52