Add a playlist selection mode
Follow the HIG "selection/edit mode" pattern by having a selection mode
for the playlist. Entering the edit mode allows us to select the songs
to be removed from the playlist without showing the "Remove" button all
the time, while still allowing a good interaction for people not using a
pointing device.

Fixes: #81