Add API for finding color distance and complementary
In order to find a color in the cover art palette that we can use for
accents, we first need to figure out the complementary color for the
dominant color in the palette. In order to do that in a convenient way,
we need to convert the RGBA color to HSLA and back.

After finding the complementary, we need to compute the distance between
that color and every component of the cover art palette we found. In
order to do that, we need to conver the RGB color palette into XYZ first
and L*a*b* second, and compute the Euclidean distance between the two
colors.