Add a progress bar when loading folders
Since folder loading can take an arbitrary amount of time, and we are
using bulk addition to avoid hammering the UI, we need to provide some
form of feedback that we're actually doing some work.

Using a thin progress bar at the top of the playlist view is as subtle
as we can be, and it has the benefit of adhering to the HIG as well.