audio: Add check for MPRIS cover art
The current implementation of the MPRIS controller in GNOME Shell will
re-use the previous cover if the art_url metadata field is set to a file
that does not exist. While this is, ostensibly, a bug in GNOME Shell, we
can work around it locally by checking if the cached cover art file
exists and is valid, before setting it as the metadata payload.