Simplify the initial state switch
We switch to the initial state whenever the queue is empty. Clearing the
queue will cause the queue to notify us that the number of songs is
zero, which will lead to switching the UI and actions to the initial
state, which in turn will cause the mode switch to the initial view.