Refactor the song state change
Simplify the property notification handler: everything that takes the
current song takes an Option<&Song>, which means we can just proxy the
State.current_song() return value as a reference.

The window title change requires a separate function, to ensure that
value reverts back to "Amberol" when there is no current song.