Make the shuffle model more predictable
The current shuffle model is a bit too flaky: shuffling will change the
current position in the queue, and unshuffling will maintain the same
position but the song will be different.

In order to make the shuffle function more predictable, we should have
the current position as an anchor: we shuffle whatever comes before (if
anything) and whatever comes next (if anything). This ensures that the
current song remains stable in the shuffled vector. The same thing
happens when unshuffling: we move the current position to the new
position of the same song.

A more predictable shuffle operation will avoid abrupt state changes,
and will keep the UI more stable.