Disable queue.clear action while adding songs
Just like we disable adding songs and folders to the playlist while
we're in the process of loading songs, we should also prevent clearing
the playlist, to avoid having to cancel the load and break the
consistency of the application state.

We might add an explicit "cancel" button to the playlist later on.

Fixes: #163