Handle nested disc folders properly
If an album directory has nested folders for each disc, we are going to
to sort by disc, which may lead to weird mixed results.

Instead of sorting using the song file and its parent folder basenames,
we use the path to the parent folder, relative to the base folder, as
the first point of comparison; for files within the same folder, we use
the song file basename.