Use the file display name as the UUID base
The basename to string conversion can fail, because file paths can
contain non-UTF-8 sequences. For the same reason, we cannot convert the
basename to an OsStr and from there to a sequence of bytes, as it would
not be portable.

Since we're using byte sequences, we might as well take the display
name, which GIO guarantees us to be UTF-8.

Fixes: #198