Implemented fuzzy searching
This implements two new types, `search::FuzzyFilter` and
`sort::FuzzySorter`, which are bound to the ListModel in window.rs.
These allow for fuzzy searching to occur and will first remove songs
without a match, and then sort them so the most relevant match is
first.

It uses the fuzzy-matcher crate to implement this.