Make sure to maintain aspect ratio of cover art pixbuf
Loading with from_stream_at_scale() and maintaining the aspect ratio
does not yield the intended results when the width and height are
identical: GdkPixbuf will favour the width, which will stretch images
that are taller than they are wider.

We need to load the image as it is to determine the aspect ratio, and
then scale it appropriately.

Fixes: #208