Stabilise shuffling behaviour
Right now we unshuffle the playlist model whenever we add new songs, if
the model was shuffled, and then shuffle it again. This makes the
playlist order absolutely wonky, and essentially only allows shuffling
once.

In order to make shuffling a bit more stable, we are going to do two
things:

- we don't shuffle songs that have already been played; the shuffle
  operation only applies to the songs that follow the current one.
- we don't shuffle songs that are added after the playlist has been
  shuffled.

Hopefully, this will make the behaviour of shuffled playlist a bit more
predictable, and won't confuse people.