Support external cover art files
If there is a "cover.png" or a "cover.jpeg" file in the same folder as
the song, then we can use that as the source of the cover art, instead
of reading the cover art from the song metadata.

We always favour the embedded cover art, because it's already in the hot
cache; loading an external file is going to take more time.

Fixes: #14