Enable next_button depending on RepeatMode
Currently, there exist three repeat modes: 'Consecutive', 'RepeatAll',
and 'RepeatOnce'. When the last song is selected, the next_button is
always set to 'disabled', even though the modes 'RepeatAll' and
'RepeatOnce' would allow the player to select the first song or restart
the currently selected song, respectively.

Fix this behavior by enabling next_button depending on the 'RepeatMode'
when the last song is selected. This also fixes the behavior of the mode
'RepeatOnce' for the last song, which was previously not working.