Don't add songs that are already in queue
Currently, songs, that are already in the queue, can still be added.
These songs will be added at the end of the queue. When selected, the
player jumps to the first song within the queue, leading to a weird
behaviour.

Fix this by not allowing duplicated songs to be added. This is checked
when the songs are about to be queued.

Fixes: https://gitlab.gnome.org/World/amberol/-/issues/248