window: Set view of playlist to the top when searching
Currently, when the search is activated and a song is being searched, the position of
the view of the playlist stays in an arbitrary position, even though the search results
are sorted with the most relevant result being at the top.

Fix this by adjusting the view of the playlist to the top which is the
most relevant result.