ui: Re-arrange the waveform scrobbler
Putting the waveform too close to the volume slider confuses things. By
placing them farther away and some additional empty space, we
de-emphasise them, and guide the eye towards the two important bits:

- the cover image
- the song details

The UI is slightly more efficient in the vertical direction as well,
instead of leaving a bunch of empty space at the top and bottom.