Always draw last bar in waveform view
Previously, the waveform view sometimes didn't draw its last bar,
depending on the exact length of the song that was playing.
Essentially, if the number of bars didn't divide the number of peaks
evenly, the condition to draw a bar would never become true for the last
one.  This commit makes sure the last bar is always drawn by checking if
the current peak is the last one.