Cope with application-less windows
During the destruction phase, when the main window is destroyed, we
might still get notifications and signals from the player object while
we wait for the window instance to disappear. If that happens, the
application's backreference on the window is going to be None, which
means we can't get the AudioPlayer instance out of it. In order to cope
with that, we can make Window::player() return an Option value, and
update the call sites to cope with that.

Fixes: #317, #319