misc: Port to AdwOverlaySplitView
This is a replacement widget of the soon-to-be-deprecated AdwFlap.
API-wise it's pretty similar to AdwFlap, so the porting was quite easy.

The thing that required a bit more changes is to adapt the app to make
use of the dynamic sidebar width which this new widget can now do. For
this reason, I had to edit some margins and paddings all around the code
and, in general, I also had to remove lots of properties that were
setting minimum sizes which we don't want anymore.